Adds initial data and filters to the app
Adds a CSV file containing initial data for the application.

Implements a data generation script to create test data if the
original CSV is not available.

Enhances the filter bar component with product, month, status,
creator, and content type filters for improved data overview
and exploration. Also adds a search input for filtering the data.

Updates styling for filter and table components.